Optimise finding shortest-distance intersect.
Instead of comparing the inner product of the two intersection,
we compare the the "t" value,
 I = P + tD
where I is the intersect vector, P is the starting position of the
ray, D is the direction of the ray.

A smaller t value represents a shorter distance travelled along
the ray, so by comparing, we may find the closed intersect.

Note that t = (I-P)/D requiring a division. We can amortise this
devision over all intersects by storing the reciprocal of D in
Ray itself. So the operation requires only one addition and one
multiplication. As opposed to 3 multiplications and 5 additions
for the dot product. If SSE or AVX operations become conveniently
available this could change.